+++ armada_backend/api_register.py
@@ -1,12 +1,16 @@
+import json
+import traceback
+
 import falcon
 
 from armada_backend import api_base, docker_client
 from armada_backend.exceptions import BadRequestException
-from armada_backend.models.services import update_service_dict
+from armada_backend.models.services import update_service_dict, save_container
 from armada_backend.models.ships import get_ship_name
 from armada_backend.utils import exists_service
+from armada_command.armada_utils import print_err
 from armada_command.consul.consul import consul_post
-from armada_command.consul.kv import kv_set, kv_remove
+from armada_command.consul.kv import kv_set, kv_remove, kv_get_recurse
 
 
 def register_service_in_consul(microservice_data):
@@ -20,8 +24,9 @@ def register_service_in_consul(microservice_data):
             'TTL': '15s',
         }
     }
-    if microservice_data['microservice_tags']:
-        consul_service_data['Tags'] = microservice_data['microservice_tags']
+    microservice_tags = microservice_data.get('microservice_tags')
+    if microservice_tags:
+        consul_service_data['Tags'] = microservice_tags
     response = consul_post('agent/service/register', consul_service_data)
     response.raise_for_status()
 
@@ -78,11 +83,21 @@ class RegisterV1(api_base.ApiCommand):
             if microservice_tags:
                 microservice_data['microservice_tags'] = microservice_tags
             microservice_version = input_json.get('microservice_version')
-            if microservice_version:
-                update_service_dict(get_ship_name(), input_json['microservice_name'], container_id,
-                                    'microservice_version', microservice_version)
+            print_err('before register_service_in_consul')
+            print_err(json.dumps(kv_get_recurse('services/'), sort_keys=True, indent=4))
             register_service_in_consul(microservice_data)
+            ship_name = get_ship_name()
+            save_container(ship_name, container_id, status='started')
+            print_err('after register_service_in_consul')
+            print_err(json.dumps(kv_get_recurse('services/'), sort_keys=True, indent=4))
+            if microservice_version:
+                print_err('updating version...')
+                update_service_dict(ship_name, input_json['microservice_name'], container_id,
+                                    'microservice_version', microservice_version)
+            print_err('after updating version')
+            print_err(json.dumps(kv_get_recurse('services/'), sort_keys=True, indent=4))
             resp.json = microservice_data
         except Exception as e:
+            traceback.print_exc()
             resp.json = {'error': 'Could not register service: {}'.format(repr(e))}
             resp.status = falcon.HTTP_400