+++ armada_backend/models/services.py
@@ -3,7 +3,8 @@ import calendar
 import time
 
 from armada_backend.models.ships import get_ship_name
-from armada_command.consul.kv import kv_get, get_env, kv_set, kv_list
+from armada_command.armada_api import get_env
+from armada_command.consul.kv import kv_get, kv_set, kv_list
 from armada_command.scripts.compat import json