+++ docker-containers/microservice/microservice/common/service_discovery.py
@@ -1,6 +1,8 @@
 from __future__ import print_function
 
 import requests
+
+from microservice.common.consul import print_err
 from microservice.defines import ARMADA_API_URL
 
 
@@ -29,6 +31,7 @@ def get_service_to_addresses():
 
 def register_service_in_armada(microservice_id, microservice_name, microservice_port, microservice_tags,
                                container_created_timestamp, single_active_instance):
+    print_err('register_service_in_armada no version')
     post_data = {
         'microservice_id': microservice_id,
         'microservice_name': microservice_name,
@@ -38,12 +41,14 @@ def register_service_in_armada(microservice_id, microservice_name, microservice_
         'single_active_instance': single_active_instance,
     }
     response = requests.post(ARMADA_API_URL + '/register', json=post_data)
+    print_err('response from old register endpoint: {} {}'.format(response.status_code, response.content))
     response.raise_for_status()
 
 
 def register_service_in_armada_v1(microservice_id, microservice_name, microservice_local_port, microservice_env,
                                   microservice_app_id, container_created_timestamp, single_active_instance,
                                   microservice_version):
+    print_err('register_service_in_armada_v1 version={}'.format(microservice_version))
     if '/' not in microservice_local_port:
         microservice_local_port = '{}/tcp'.format(microservice_local_port)
     post_data = {
@@ -57,6 +62,7 @@ def register_service_in_armada_v1(microservice_id, microservice_name, microservi
     }
     url = '{}/v1/register/{}'.format(ARMADA_API_URL, microservice_id)
     response = requests.post(url, json=post_data)
+    print_err('response from new register endpoint: {} {}'.format(response.status_code, response.content))
     if response.status_code == 404:
         raise UnsupportedArmadaApiException('Endpoint /v1/register is unavailable.')
     response.raise_for_status()